+++ src/com/skinnybonesarv/beta/FileManagement.java
@@ -1,0 +1,13 @@
+package com.skinnybonesarv.beta;
+
+import java.util.List;
+
+public class FileManagement {
+
+    public static void addToFile(List<Student> students) {
+        addToFile(students);
+        // Add contents of List to file
+
+    }
+
+}
