+++ src/com/skinnybonesarv/beta/Main.java
@@ -1,0 +1,11 @@
+package com.skinnybonesarv.beta;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Welcome to SkinnyBonesArv's School Management System!");
+        // Welcome Screen
+        do {
+            // Do something
+        } while (true);
+    }
+}
