+++ src/com/skinnybonesarv/beta/StudentManagment.java
@@ -1,0 +1,14 @@
+package com.skinnybonesarv.beta;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class StudentManagment {
+
+    private List<Student> students = new ArrayList<>();
+
+    public void addStudent() {
+        FileManagement.addToFile(students);
+    }
+
+}
