--- src/com/test/FileManagement.java
@@ -1,13 +1,0 @@
-package com.test;
-
-import java.util.List;
-
-public class FileManagement {
-
-    public static void addToFile(List<Student> students) {
-        addToFile(students);
-        // Add contents of List to file
-
-    }
-
-}
