--- src/com/test/Main.java
@@ -1,11 +1,0 @@
-package com.test;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Welcome to SkinnyBonesArv's School Management System!");
-        // Welcome Screen
-        do {
-            // Do something
-        } while (true);
-    }
-}
